backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-rest-auth==7.0.1
django-allauth==65.10.0
django-filter==24.1
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
